Remove disabled-look from days in previous/next month
While working with the react-datepicker I discovered that days in the
previous and/or next months look like they're disabled, but they're
not.

We should make it more clear dates in the previous and/or next month
are selectable. In order to achieve this, I propose to remove this
`this-month` state and make the day's color default to black.